Move SourceProviderPass to "Compiler" sub-namespace
